Rename transCoordsUsingJoints to transCoords
