+++ programs/CMakeLists.txt
@@ -2,4 +2,4 @@
 
 add_subdirectory(keyboardController)
 add_subdirectory(streamingDeviceController)
-add_subdirectory(transCoordsUsingJoints)
+add_subdirectory(transCoords)
